fix: full user implemented

+++ main-service/src/main/java/ru/practicum/user/service/UserServiceImpl.java
@@ -1,6 +1,9 @@
 package ru.practicum.user.service;
 
+import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -10,10 +13,12 @@ import ru.practicum.user.mappers.UserMapper;
 import ru.practicum.user.model.User;
 import ru.practicum.user.repository.UserRepository;
 
+import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
 
 @Service
+@Slf4j
 @RequiredArgsConstructor
 public class UserServiceImpl implements UserService {
 
@@ -29,22 +34,33 @@ public class UserServiceImpl implements UserService {
     @Override
     @Transactional
     public void deleteUser(Long userId) {
+        if (!userRepository.existsById(userId)) {
+            throw new EntityNotFoundException("User not found with id: " + userId);
+        }
         userRepository.deleteById(userId);
     }
 
     @Override
     @Transactional
-    public List<UserDto> getUsers(List<Long> ids, int from, int size) {
-        Pageable pageable = PageRequest.of(from, size);
+    public List<UserDto> getUsers(String ids, int from, int size) {
+        List<Long> idList = null;
 
-        List<User> users;
-        if (ids != null && !ids.isEmpty()) {
-            users = userRepository.findAllById(ids);
-        } else {
-            users = userRepository.findAll(pageable).getContent();
+        if (!ids.equals("null")) {
+            idList = Arrays.stream(ids.split(","))
+                    .map(Long::parseLong)
+                    .collect(Collectors.toList());
         }
 
-        return users.stream()
+        Pageable pageable = PageRequest.of(from, size);
+        Page<User> users;
+
+        if (idList != null && !idList.isEmpty()) {
+            users = userRepository.findAllByIdIn(idList, pageable);
+        } else {
+            users = userRepository.findAll(pageable);
+        }
+
+        return users.getContent().stream()
                 .map(UserMapper::mapToUserDto)
                 .collect(Collectors.toList());
     }
